+++ packages/scaffolder-core/jest.config.js
@@ -1,0 +1,4 @@
+const baseJestConfig = require('../../jest.base.config.js');
+module.exports = {
+	...baseJestConfig
+};