+++ packages/scaffolder-core/CHANGELOG.md
@@ -5,7 +5,15 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [1.6.20](https://github.com/galElmalah/ctf/compare/v0.0.1...v1.6.20) (2020-10-15)
 
-**Note:** Version bump only for package root
+**Note:** Version bump only for package scaffolder-core
+
+
+
+
+
+## [1.6.19](https://github.com/galElmalah/ctf/compare/v0.0.1...v1.6.19) (2020-10-15)
+
+**Note:** Version bump only for package scaffolder-core
 
 
 
@@ -13,4 +21,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.0.1](https://github.com/galElmalah/ctf/compare/v1.6.18...v0.0.1) (2020-10-15)
 
-**Note:** Version bump only for package root
+**Note:** Version bump only for package @scaffolder-org/core